ENH: Remove conditional code for old VTK
This commit removes some conditional coding having to do with the use of
an earlier version of VTK. The main issue is that VTK determines its
version by the use of git tree, which is not available when building VTK
from a source package. Therefore this conditional won't work on the
Gentoo package.