ENH: Add additional module paths by drag-and-drop to application
Python scripted modules folders can now be added to "Additional module paths" by drag-and-dropping module .py files or folders to the application window.

Make the application window a parent of the reader selector popup to make it easier to raise the popup to the top (otherwise the tiny reader selector popup may get stuck below the application icon where the files were drag-and-dropped from).

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>