WIP: Enable offline build of Slicer and handle local python wheel files
This commit modifies the SuperBuild/External_python-pip.cmake and SuperBuild/External_python-setuptools.cmake files to allow for the offline build of Slicer. Specifically, the changes enable the handling of local python wheel files instead of using remote pull.

The changes include the addition of a block that considers the possibility of dependencies being given as a single file or a directory of dependencies. In the case of a directory, all the *.whl files in the directory will be installed.